Start of read_only feature for users

+++ app/controllers/clients_controller.rb
@@ -4,9 +4,15 @@ class ClientsController < ApplicationController
   end
 
   def destroy
-    resp = Client.destroy(params[:id])
-    respond_to do |format|
-      format.json { render :json => {:data => (resp == 202).to_s}.to_json }
+    if current_user.read_only?
+      respond_to do |format|
+        format.json { render :json => false }
+      end
+    else
+      resp = Client.destroy(params[:id])
+      respond_to do |format|
+        format.json { render :json => {:data => (resp == 202).to_s}.to_json }
+      end
     end
   end
 end
